# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Starwars",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/starships": {
            "get": {
                "tags": [
                    "Starship"
                ],
                "summary": "Query all starship",
                "operationId": "indexStarship",
                "responses": {
                    "200": {
                        "description": "returns the registered starships"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Starship"
                ],
                "summary": "register starship, only place 6 fields but 18 can be registered, this is just an example",
                "operationId": "storeStarship",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "manufacturer": {
                                        "type": "string"
                                    },
                                    "cost_in_credits": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "string"
                                    },
                                    "max_atmosphering_speed": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jedi Angel based",
                                    "model": "Republic Assault ship",
                                    "manufacturer": "Rendili StarDrive",
                                    "cost_in_credits": "125000000",
                                    "length": "1088",
                                    "max_atmosphering_speed": 1050
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns the registered object"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/starships/{id}": {
            "get": {
                "tags": [
                    "Starship"
                ],
                "summary": "Return a query",
                "operationId": "showStarship",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a vehicle query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Starship"
                ],
                "summary": "updated starship, only place 6 fields but 18 can be updated, this is just an example",
                "operationId": "updateStarship",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "manufacturer": {
                                        "type": "string"
                                    },
                                    "cost_in_credits": {
                                        "type": "string"
                                    },
                                    "length": {
                                        "type": "string"
                                    },
                                    "max_atmosphering_speed": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Jedi Angel based",
                                    "model": "Republic Assault ship",
                                    "manufacturer": "Rendili StarDrive",
                                    "cost_in_credits": "125000000",
                                    "length": "1088",
                                    "max_atmosphering_speed": 1050
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns the updated object"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Starship"
                ],
                "summary": "Delete to data",
                "operationId": "destroyStarship",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a delete query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Return a query",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a registered query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a data",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Angel",
                                    "email": "angelmserranog@gmail.com",
                                    "password": 12345678
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "returns a updated query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete to data",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns a delete query"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "500": {
                        "description": "internal error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}